Changed pages to hooks and components and also upgraded UI. Made repository easier to read
