+++ apps/client/src/components/LoginView.tsx
@@ -1,0 +1,127 @@
+import React from 'react';
+
+interface LoginViewProps {
+  email: string;
+  setEmail: (email: string) => void;
+  password: string;
+  setPassword: (password: string) => void;
+  message: string;
+  isLoading: boolean;
+  handleLogin: (e: React.FormEvent) => Promise<void>;
+  navigateToRegister: () => void;
+}
+
+const LoginView: React.FC<LoginViewProps> = ({
+  email,
+  setEmail,
+  password,
+  setPassword,
+  message,
+  isLoading,
+  handleLogin,
+  navigateToRegister
+}) => {
+  return (
+    <div className="flex items-center justify-center min-h-screen px-4 py-12">
+      <div className="w-full max-w-md">
+        <div className="text-center mb-8">
+          <h1 className="text-3xl font-bold text-gray-800">Welcome Back</h1>
+          <p className="text-gray-600 mt-2">Sign in to continue to your account</p>
+        </div>
+        
+        <form 
+          className="bg-white p-8 rounded-lg shadow-lg border border-gray-100" 
+          onSubmit={handleLogin}
+        >
+          <h2 className="text-2xl font-semibold text-gray-800 mb-6 text-center">Login</h2>
+          
+          <div className="mb-4">
+            <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">Email Address</label>
+            <div className="relative">
+              <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                <svg className="h-5 w-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                  <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z" />
+                  <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z" />
+                </svg>
+              </div>
+              <input
+                id="email"
+                className="p-3 pl-10 w-full border border-gray-300 rounded-md focus:ring-blue-500 focus:border-blue-500 transition-colors text-base"
+                type="email"
+                placeholder="Enter your email"
+                value={email}
+                onChange={e => setEmail(e.target.value)}
+                required
+              />
+            </div>
+          </div>
+          
+          <div className="mb-6">
+            <div className="flex items-center justify-between mb-1">
+              <label htmlFor="password" className="block text-sm font-medium text-gray-700">Password</label>
+              <a href="#" className="text-sm text-blue-600 hover:text-blue-800 transition-colors">
+                Forgot password?
+              </a>
+            </div>
+            <div className="relative">
+              <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                <svg className="h-5 w-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                  <path fillRule="evenodd" d="M5 9V7a5 5 0 0110 0v2a2 2 0 012 2v5a2 2 0 01-2 2H5a2 2 0 01-2-2v-5a2 2 0 012-2zm8-2v2H7V7a3 3 0 016 0z" clipRule="evenodd" />
+                </svg>
+              </div>
+              <input
+                id="password"
+                className="p-3 pl-10 w-full border border-gray-300 rounded-md focus:ring-blue-500 focus:border-blue-500 transition-colors text-base"
+                type="password"
+                placeholder="Enter your password"
+                value={password}
+                onChange={e => setPassword(e.target.value)}
+                required
+              />
+            </div>
+          </div>
+          
+          {message && (
+            <div className="mb-4 p-3 bg-red-50 border border-red-200 text-red-600 rounded-md flex items-center">
+              <svg className="h-5 w-5 mr-2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+                <path fillRule="evenodd" d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-7 4a1 1 0 11-2 0 1 1 0 012 0zm-1-9a1 1 0 00-1 1v4a1 1 0 102 0V6a1 1 0 00-1-1z" clipRule="evenodd" />
+              </svg>
+              {message}
+            </div>
+          )}
+          
+          <button 
+            className="w-full p-3 rounded-md bg-blue-600 hover:bg-blue-700 text-white font-medium transition-colors focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 text-base"
+            type="submit"
+            disabled={isLoading}
+          >
+            {isLoading ? 'Signing In...' : 'Sign In'}
+          </button>
+          
+          <div className="mt-4 relative">
+            <div className="absolute inset-0 flex items-center">
+              <div className="w-full border-t border-gray-200"></div>
+            </div>
+            <div className="relative flex justify-center text-sm">
+              <span className="px-2 bg-white text-gray-500">Or</span>
+            </div>
+          </div>
+          
+          <button 
+            className="mt-4 w-full p-3 rounded-md bg-gray-100 hover:bg-gray-200 text-gray-800 font-medium transition-colors border border-gray-300 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 text-base flex items-center justify-center"
+            type="button" 
+            onClick={navigateToRegister}
+          >
+            Create New Account
+          </button>
+        </form>
+        
+        <p className="mt-6 text-center text-sm text-gray-600">
+          Need help? <a href="#" className="font-medium text-blue-600 hover:text-blue-500">Contact Support</a>
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default LoginView;
